Fixed various bugs.

+++ src/dgraph.c
@@ -457,9 +457,9 @@ static void add_edge(Graph *graph, Wire wire, const char *filePath) {
         *(graph->wires) = wire;
     } else {
         // Use binary insertion, since the list should be sorted!
-        size_t left   = 0;
-        size_t right  = graph->numWires - 1;
-        size_t middle = (left + right) / 2;
+        int left   = 0;
+        int right  = (int)graph->numWires - 1;
+        int middle = (left + right) / 2;
 
         while (left <= right) {
             middle = (left + right) / 2;
@@ -483,7 +483,7 @@ static void add_edge(Graph *graph, Wire wire, const char *filePath) {
         graph->wires =
             (Wire *)d_realloc(graph->wires, graph->numWires * sizeof(Wire));
 
-        if (middle < graph->numWires - 1) {
+        if (middle < (int)graph->numWires - 1) {
             memmove(graph->wires + middle + 1, graph->wires + middle,
                     (graph->numWires - middle - 1) * sizeof(Wire));
         }
